use unshare instead of chroot

+++ cluster.go
@@ -4,7 +4,9 @@ import (
 	"context"
 	"errors"
 	"os"
+	"os/exec"
 	"sync"
+	"syscall"
 
 	"github.com/cybozu-go/cmd"
 	"github.com/cybozu-go/log"
@@ -97,8 +99,7 @@ func (c *Cluster) Cleanup(r *Runtime) error {
 	}
 
 	CleanupNetworks(r, c)
-
-	return CleanupRootfs()
+	return nil
 }
 
 // GetNetwork looks up the network by name.
@@ -151,9 +152,58 @@ func (c *Cluster) GetPod(name string) (*Pod, error) {
 	return p, nil
 }
 
+func umount(mp string) error {
+	return exec.Command("umount", mp).Run()
+}
+
+func mount(fs, dest, options string) error {
+	err := os.MkdirAll(dest, 0755)
+	if err != nil {
+		return err
+	}
+	log.Info("mount", map[string]interface{}{
+		"fs":   fs,
+		"dest": dest,
+	})
+	c := exec.Command("mount", "-t", fs, "-o", options, fs, dest)
+	c.Stderr = os.Stderr
+	c.Stdout = os.Stdout
+	return c.Run()
+}
+
+func (c *Cluster) startPodOnly(ctx context.Context, r *Runtime) error {
+	err := mount("tmpfs", "/run", "rw")
+	if err != nil {
+		return err
+	}
+	defer umount("/run")
+
+	for _, p := range c.Pods {
+		err := p.Prepare(ctx)
+		if err != nil {
+			return err
+		}
+	}
+
+	env := cmd.NewEnvironment(ctx)
+	for _, p := range c.Pods {
+		p := p
+		env.Go(func(ctx context.Context) error {
+			return p.Start(ctx, r)
+		})
+	}
+	env.Stop()
+
+	return env.Wait()
+}
+
 // Start constructs the virtual data center with given resources.
 // It stop when ctx is cancelled.
 func (c *Cluster) Start(ctx context.Context, r *Runtime) error {
+	if r.podOnly {
+		return c.startPodOnly(ctx, r)
+	}
+
 	defer os.RemoveAll(r.tempDir)
 
 	if r.force {
@@ -163,13 +213,7 @@ func (c *Cluster) Start(ctx context.Context, r *Runtime) error {
 		}
 	}
 
-	root, err := NewRootfs()
-	if err != nil {
-		return err
-	}
-	defer root.Destroy()
-
-	err = createNatRules()
+	err := createNatRules()
 	if err != nil {
 		return err
 	}
@@ -199,13 +243,6 @@ func (c *Cluster) Start(ctx context.Context, r *Runtime) error {
 			"name": img.Name,
 		})
 		err := img.Prepare(ctx, r.imageCache)
-		if err != nil {
-			return err
-		}
-	}
-
-	for _, p := range c.Pods {
-		err := p.Prepare(ctx)
 		if err != nil {
 			return err
 		}
@@ -246,16 +283,23 @@ func (c *Cluster) Start(ctx context.Context, r *Runtime) error {
 
 	env = cmd.NewEnvironment(ctx)
 	env.Go(bmcServer.handleNode)
-	for _, p := range c.Pods {
-		p := p
-		env.Go(func(ctx context.Context) error {
-			return p.Start(ctx, r, root.Path())
-		})
-	}
 	for _, vm := range vms {
 		vm := vm
 		env.Go(func(ctx context.Context) error {
 			return vm.cmd.Wait()
+		})
+	}
+
+	if len(c.Pods) > 0 {
+		c := exec.Command("/proc/self/exe", append([]string{"--pod-only"}, os.Args[1:]...)...)
+		c.Stdin = os.Stdin
+		c.Stdout = os.Stdout
+		c.Stderr = os.Stderr
+		c.SysProcAttr = &syscall.SysProcAttr{
+			Unshareflags: syscall.CLONE_NEWNS,
+		}
+		env.Go(func(ctx context.Context) error {
+			return c.Run()
 		})
 	}
 	env.Stop()
